+++ dds_cli/__main__.py
@@ -9,6 +9,7 @@ import concurrent.futures
 import itertools
 import logging
 import os
+from re import T
 import sys
 
 # Installed
@@ -24,6 +25,7 @@ import questionary
 # Own modules
 import dds_cli
 import dds_cli.account_manager
+import dds_cli.unit_manager
 import dds_cli.data_getter
 import dds_cli.data_lister
 import dds_cli.data_putter
@@ -159,6 +161,17 @@ def dds_main(click_ctx, verbose, log_file, no_prompt, token_path):
 @project_option(required=False)
 @sort_projects_option()
 @folder_option(help_message="List contents of this project folder.")
+@click.option(
+    "--binary",
+    "-b",
+    required=False,
+    is_flag=True,
+    default=False,
+    help=(
+        "Use binary unit prefixes (e.g. KiB instead of KB, "
+        "MiB instead of MB) for size and usage columns."
+    ),
+)
 # Flags
 @json_flag(help_message="Output in JSON format.")
 @size_flag(help_message="Show size of project contents.")
@@ -168,7 +181,7 @@ def dds_main(click_ctx, verbose, log_file, no_prompt, token_path):
 @click.option("--projects", "-lp", is_flag=True, help="List all project connected to your account.")
 @click.pass_obj
 def list_projects_and_contents(
-    click_ctx, project, folder, sort, json, size, tree, usage, users, projects
+    click_ctx, project, folder, sort, json, size, tree, usage, binary, users, projects
 ):
     """List the projects you have access to or the project contents.
 
@@ -186,6 +199,7 @@ def list_projects_and_contents(
                 no_prompt=click_ctx.get("NO_PROMPT", False),
                 json=json,
                 token_path=click_ctx.get("TOKEN_PATH"),
+                binary=binary,
             ) as lister:
                 projects = lister.list_projects(sort_by=sort)
                 if json:
@@ -333,7 +347,7 @@ def login(click_ctx):
     try:
         with dds_cli.auth.Auth(token_path=click_ctx.get("TOKEN_PATH")):
             # Authentication token renewed in the init method.
-            LOG.info("[green] :white_check_mark: Authentication token renewed![/green]")
+            LOG.info("[green] :white_check_mark: Authentication token created![/green]")
     except (
         dds_cli.exceptions.APIError,
         dds_cli.exceptions.AuthenticationError,
@@ -402,16 +416,24 @@ def user_group_command(_):
 # ************************************************************************************************ #
 
 # -- dds user ls -- #
+# TODO: Move this to dds unit?
 @user_group_command.command(name="ls")
+@click.option(
+    "--unit",
+    "-u",
+    required=False,
+    type=str,
+    help="Super Admins only: The unit which you wish to list the users in.",
+)
 @click.pass_obj
-def list_users(click_ctx):
+def list_users(click_ctx, unit):
     """List users."""
     try:
         with dds_cli.account_manager.AccountManager(
             no_prompt=click_ctx.get("NO_PROMPT", False),
             token_path=click_ctx.get("TOKEN_PATH"),
         ) as lister:
-            lister.list_unit_users()
+            lister.list_unit_users(unit=unit)
     except (
         dds_cli.exceptions.AuthenticationError,
         dds_cli.exceptions.ApiResponseError,
@@ -881,8 +903,16 @@ def retract_project(click_ctx, project):
 @project_status.command(name="archive", no_args_is_help=True)
 # Options
 @project_option(required=True)
+# Flags
+@click.option(
+    "--abort",
+    required=False,
+    is_flag=True,
+    default=False,
+    help="Something has one wrong in the project.",
+)
 @click.pass_obj
-def archive_project(click_ctx, project: str):
+def archive_project(click_ctx, project: str, abort: bool = False):
     """Manually archive a released project.
 
     This deletes all project data.
@@ -899,7 +929,7 @@ def archive_project(click_ctx, project: str):
                 no_prompt=click_ctx.get("NO_PROMPT", False),
                 token_path=click_ctx.get("TOKEN_PATH"),
             ) as updater:
-                updater.update_status(new_status="Archived")
+                updater.update_status(new_status="Archived", is_aborted=abort)
         except (
             dds_cli.exceptions.APIError,
             dds_cli.exceptions.AuthenticationError,
@@ -933,39 +963,6 @@ def delete_project(click_ctx, project: str):
                 token_path=click_ctx.get("TOKEN_PATH"),
             ) as updater:
                 updater.update_status(new_status="Deleted")
-        except (
-            dds_cli.exceptions.APIError,
-            dds_cli.exceptions.AuthenticationError,
-            dds_cli.exceptions.DDSCLIException,
-            dds_cli.exceptions.ApiResponseError,
-        ) as err:
-            LOG.error(err)
-            sys.exit(1)
-
-
-# -- dds project status abort -- #
-@project_status.command(name="abort", no_args_is_help=True)
-# Options
-@project_option(required=True)
-@click.pass_obj
-def abort_project(click_ctx, project: str):
-    """Abort a released project.
-
-    This deletes all project data.
-    """
-    proceed_deletion = (
-        True
-        if click_ctx.get("NO_PROMPT", False)
-        else dds_cli.utils.get_deletion_confirmation(action="abort", project=project)
-    )
-    if proceed_deletion:
-        try:
-            with dds_cli.project_status.ProjectStatusManager(
-                project=project,
-                no_prompt=click_ctx.get("NO_PROMPT", False),
-                token_path=click_ctx.get("TOKEN_PATH"),
-            ) as updater:
-                updater.update_status(new_status="Archived", is_aborted=True)
         except (
             dds_cli.exceptions.APIError,
             dds_cli.exceptions.AuthenticationError,
@@ -1405,16 +1402,58 @@ def rm_data(click_ctx, project, file, folder, rm_all):
             if rm_all:
                 remover.remove_all()
 
-            elif file:
-                remover.remove_file(files=file)
+            else:
+                if file:
+                    remover.remove_file(files=file)
 
-            elif folder:
-                remover.remove_folder(folder=folder)
+                if folder:
+                    remover.remove_folder(folder=folder)
     except (
         dds_cli.exceptions.AuthenticationError,
         dds_cli.exceptions.APIError,
         dds_cli.exceptions.DDSCLIException,
         dds_cli.exceptions.ApiResponseError,
+    ) as err:
+        LOG.error(err)
+        sys.exit(1)
+
+
+####################################################################################################
+####################################################################################################
+## UNIT #################################################################################### UNIT ##
+####################################################################################################
+####################################################################################################
+
+
+@dds_main.group(name="unit", no_args_is_help=True)
+@click.pass_obj
+def unit_group_command(_):
+    """Group command for managing units.
+
+    Only for Super Admins.
+    """
+
+
+# ************************************************************************************************ #
+# UNIT COMMANDS ******************************************************************** UNIT COMMANDS #
+# ************************************************************************************************ #
+
+# -- dds unit ls -- #
+@unit_group_command.command(name="ls", no_args_is_help=False)
+@click.pass_obj
+def list_units(click_ctx):
+    """List all units."""
+    try:
+        with dds_cli.unit_manager.UnitManager(
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
+        ) as lister:
+            lister.list_all_units()
+    except (
+        dds_cli.exceptions.AuthenticationError,
+        dds_cli.exceptions.ApiResponseError,
+        dds_cli.exceptions.ApiRequestError,
+        dds_cli.exceptions.DDSCLIException,
     ) as err:
         LOG.error(err)
         sys.exit(1)