+++ dds_cli/user.py
@@ -13,6 +13,7 @@ import pathlib
 import requests
 import simplejson
 import stat
+import subprocess
 
 # Installed
 import rich
@@ -80,7 +81,7 @@ class User:
                     "No saved token found, or token has expired, proceeding with authentication"
                 )
             else:
-                LOG.info("Attempting to renew the session token")
+                LOG.info("Attempting to create the session token")
             self.token = self.__authenticate_user()
             token_file.save_token(self.token)
 
@@ -196,7 +197,7 @@ class User:
         username = None
         if tokenfile.file_exists():
             token = tokenfile.read_token()
-            if not tokenfile.token_expired(token=token):
+            if token and not tokenfile.token_expired(token=token):
                 try:
                     response = requests.get(
                         dds_cli.DDSEndpoint.DISPLAY_USER_INFO,
@@ -259,6 +260,22 @@ class TokenFile:
         with self.token_file.open("w") as file:
             file.write(token)
 
+        if os.name == "nt":
+            cli_username = os.environ.get("USERNAME")
+            try:
+                subprocess.check_call(
+                    [
+                        "icacls.exe",
+                        str(self.token_file),
+                        "/inheritance:r",
+                        "/grant",
+                        f"{cli_username}:(R,W)",
+                    ],
+                    stdout=subprocess.DEVNULL,
+                    stderr=subprocess.DEVNULL,
+                )
+            except subprocess.CalledProcessError as exc:
+                LOG.error("Failed to set token file permissions")
         LOG.debug("New token saved to file.")
 
     def delete_token(self):
@@ -273,13 +290,16 @@ class TokenFile:
 
         Returns None otherwise.
         """
-        st_mode = os.stat(self.token_file).st_mode
-        permissions_octal = oct(stat.S_IMODE(st_mode))
-        permissions_readable = stat.filemode(st_mode)
-        if permissions_octal != "0o600":
-            raise exceptions.DDSCLIException(
-                message=f"Token file permissions are not properly set, (got {permissions_readable} instead of required '-rw-------'). Please remove {self.token_file} and rerun the command."
-            )
+        if os.name != "nt":
+            st_mode = os.stat(self.token_file).st_mode
+            permissions_octal = oct(stat.S_IMODE(st_mode))
+            permissions_readable = stat.filemode(st_mode)
+            if permissions_octal != "0o600":
+                raise exceptions.DDSCLIException(
+                    message=f"Token file permissions are not properly set, (got {permissions_readable} instead of required '-rw-------'). Please remove {self.token_file} and rerun the command."
+                )
+        else:
+            LOG.info("Unable to confirm whether file permissions are correct on Windows.")
 
     def token_expired(self, token):
         """Check if the token has expired or is about to expire soon based on the UTC time.